+++ include/delfem2/femstokes.h
@@ -188,7 +188,7 @@ void MergeLinSys_Stokes3D_Static(
     const std::vector<double>& aVal,
     const std::vector<double>& aVelo)
 {
-  const unsigned int np = aXYZ.size()/3;
+  const unsigned int np = static_cast<unsigned int>(aXYZ.size()/3);
   const unsigned int nDoF = np*4;
   //
   mat_A.setZero();
